requirements.txt
flask==2.3.3
flask-cors==4.0.0
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2 